ActionViewHelper made to work with/without html_safe.
Verified that tests are passing with ActiveSupport and ActionPack versions:
* 3.0.0.rc
* 2.3.8
* 2.3.2

I removed the support for ActionPack < 2.2 in the ActionViewHelper. As far as I can
tell, since we have a dependency on ActiveSupport >= 2.3.2 and ActionPack always has
a hard dependency on its matching version of ActiveSupport, there is no way that someone
could be using this version of ActiveMerchant along with ActionPack < 2.3.2.