Adds CCV code support for Authorize.net CIM transactions. Since CCV codes are not stored as part of the payment profile they must be supplied with any transaction (including validateCustomerPaymentProfile) that wishes to use CCV.
Updates the Authorize.net CIM tests to use prior_auth_capture instead of capture_only in the auth then capture unit test. This is a better/more correct example as capture only is not supposed to be used with auth_only.  Rather it is intended to be used with transations that were authorized through some other system or required voice authorization.