Braintree Blue: Make address country map to alpha2
The (undocumented) standard for the :country key in addresses is
for it to be the standard two-character code for the country.
The old behavior for Braintree Blue gateway, fixed in this commit,
was to take that field and map it to :country_name in Braintree,
which generally didn't work as expected.

Now the gateway puts :country into the :country_code_alpha2 field,
and also allows passing :country_code_alpha2, :country_code_alpha3,
and :country_code_numeric if more control is needed.