Wirecard gateway: Fix for missing address hash
If the address has was not passed in, the gateway would throw an error
when attempting to pull the country information out of the non-existent
hash.

Fixed via a general cleanup of handling the options hash; previously,
the @options instance variable was overwritten with call specific
options. Now the @options is merged with the passed in options, and
the instance variable isn't ever touched.

Also did general formatting cleanup, and refactored how the request is
built.