Braintree Blue gateway: Add detail to responses
This change is to enable the use of Braintree's client side encryption.
When the CC number is encrypted, you can't figure out what type it is or
what the last four digits are. Braintree's response tells you though.
This change is just adding those parameters on the respective hashes for
transaction and customer create calls.

Closes #409.