Plug'n Pay gateway: Add tests for partial capture
Partial captures were not being completed through Plug'n Pay, and
initially it looked like a code change was needed to fix it (using
reauth). However, it turned out to be a configuration problem with the
account on the Plug'n Pay side, so this commit simply adds tests
specifically for partial capture.

Closes #410.