Ogone gateway: loosen up signature warnings
Previously, deprecation warnings were being spit out anytime a
transaction was run without a signature, and anytime a signature
strength less than 512 was used. This was a bit onerous, as it required
not just code changes, but configuration changes for all legacy Ogone
accounts. While Ogone may force all customers to start using signatures
in the future, in all likelihood they'll just make it the default for
new accounts (as they do today).

This commit does two things. First, it continues to give a deprecation
warning if no signature is supplied, but allows that warning to be
overridden by simply providing a signature_encryptor of "none". Second,
it dispenses with any warning regarding signature strength, since that
is still fully configurable in the Ogone backend.