Add refund method to Viaklix gateway.
This brings Viaklix and Elavon in line with the other gateways in that
the #credit method is deprecated and the refund method is recommended
path.

Viaklix and Elavon are a bit unconventional in that they don't seem to
support credit via reference.  Instead, they require the full card
details again to do a credit.  In fact, they seem to require the card to
do a capture as well.

All in all, it's sub-optimal, but at least things are now a little more
consistent with the other direct gateways.