SagePay integration: Remove truncate dependency
Using ActiveSupport's String#truncate is problematic, since it is not
present in ActiveSupport 2.3.x, which ActiveMerchant still tries to
provide support for. It is also problematic in that it adds yet another
ActiveSupport dependency, when I think we'd prefer to strip out
ActiveSupport usage altogether.

This makes the description passed via the SagePay Form integration a bit
less pretty, but truncate can easily be applied by callers before
handing the description off to the helper if the old behavior is still
desired.