Make the default Gemfile Rails dependency explicit
Tests in 1.8 started failing recently on Travis, and I tracked it back
to a change in the Rails/ActiveSupport version being loaded. We did not
change anything in our dependencies, but I think something has changed
in rubygems.org that has triggered an older version (2.3.14) to be
loaded, whereas 2.3.14 was loaded previously.

If we want to continue to claim 2.3.x support, we need to set up a
proper Gemfile matrix and fix the tests when running on 2.3.x. Or we
could just drop 2.3.x.