Fix tests dependent on empty tag generation
Newer versions of Builder (which may or may not be selected by bundler)
generate an auto-closed XML tag in the case of an empty tag rather than
an open and a close tag. These tests were depending on the latter
behavior.

These tests are way too brittle for my taste, but the goal at this point
is to simply get all tests passing again.