Mercury gateway: Fix authorizations API
The original implementation required passing around a ton of extra
fields for follow-on transactions, rather than encoding those values
into the authorization. Note however, that even with this changed, all
Mercury transactions appear to require the full credit card to be passed
in, which is non-standard ActiveMerchant.

Also strips out all support for track2, voice auths, and batch
processing, since these are not typically supported by ActiveMerchant.
If a champion that is going to use them actively wants to add them back
in, be my guest :-)