PaymentExpress: Add support for optional fields
Even though not documented for PxPost (it is documented for WebService),
the ClientType element on transactions can affect the way PaymentExpress
submits transaction to the acquirers. This commit gives control over
that element through the `:client_type` element in the options hash.

For completeness, access has also been given to the remaining optional
elements: `TxnData1`, `TxnData2`, and `TxnData3`, which can be set via
`:txn_data1`, `:txn_data2`, and `:txn_data2`, respectively, in the
options hash.

Closes #511.