Refactor Gateway#test? and assignment of @options
Up till now, we've depended on each gateway defining a sane #test?
method, which ended up with almost every gateway having the exact same
definition of #test?. This commit pulls that common definition up to the
Gateway class.

As part of doing this, since the definition of #test? is dependent on
@options, I had to also pull up `@options = options` to
Gateway#initialize, which also eliminated a ton of duplication between
gateways.

This should be a strict refactoring - no functionality or external APIs
should be affected.

Closes #408.