Orbital gateway: Fix infinite connection retry
Orbital could very easily land in an infinite retry situation if unable
to connect. In general retry should be a concern external to
ActiveMerchant, but in the Orbital case there is a secondary url that
should be tried before giving up. This removes generalized retries, but
retains a single retry at the secondary url.

Closes #464.