Litle: Add support for tokenization
Adds support for passing a credit card OR token to the authorize and
purchase methods.

Note: a token is not a reference (authorization <litleTxnId>) to a
previous transaction, but a token representing a credit card. A token is
created by the gateway (if configured for tokenization) and returned in
either explicit or implicit register token requests.

Adds certification tests defined by the LitleXML Reference Guide section
2.4.4.

Starts refactoring the LitleGateway class to match the patterns used by
other gateways.

Closes #514.