Braintree: Store sets vault id as authorization
For consistency's sake, it's helpful to always have the result of an
explicit Gateway#store call set the Response#authorization to be the
assigned vault id (though it's fine to also have it in a gateway-
specific location as well). This implements this for both Braintree
gateway API's, including unit and remote tests.