NetPay: Improve support for void
This makes the interface to #void consistent with the other gateways.
It does this by building the authorization and then splitting it later
to get the parts that we need.

This is similar to how we're handling void in the Redsys gateway which
also seems to need an amount on their call to void.