NetPay: Fix the signature for void
This makes the interface to #void consistent with the other gateways.
It does this by building the authorization and then splitting it later
to get the parts that we need. Passing the amount to void is still
supported via options[:amount].

No deprecation is being done for the incorrect signature since Netpay
was added so recently.

Also fixes an intermittent test failure on 1.8 caused by Hash order
dependence.

Closes #553.