Individual gateways work with :test => false
If Base.test? was true but a developer attempted to override it by
specifying :test => false in an individual gateway's options, it
continued to operate in test mode due to interaction of the false value
with the || operator being used. This instead always uses the gateway's
:test option if it is set in the options hash.

Reported by @DiegoSalazar. Fixes #586.