Remove auth_reversal from the public interface and implement it in #void
To maintain the same interface as the other gateways #void now detects what kind of transaction it is voiding based on an indentification that is composed of the transaction-id and the kind of transaction.  If it is voiding an authorization it will perform an auth_reversal, otherwise it will perform a void. This means Response.authorization will also now always contain this composed identification.