Swap out REXML for Nokogiri in payflow common parser.
REXML is very strict with its xml parsing so fails hard on invalid
characters, Nokogiri strips them and continues on. This is a much better
experience for ActiveMerchant users.

I've only replaced the parser in Payflow common, but we should work on
moving everything to Nokogiri since it is a much better XML parser.