Fat Zebra: More consistent handling of tokens
Switches to using the ActiveMerchant standard of passing in the token as
the payment method rather than using a Hash. The CVV can still be passed
via the options hash. The old method of passing in a Hash is deprecated.

Also cleans up the error handling to be more robust.

Closes #715.