On March 22, 2012, Moneris implemented a method for reversing a preauthorization to comply with updated standards. From their announcement which was emailed to merchants:
"Within eSelectPlus to perform a Preauthorization Reversal for 100% of the original Preauth amount please perform a Capture transaction with an amount of 0.00.  This will cause the preauthorization to be fully reversed.  For transactions that are only completed for less than the Preauth amount, the Moneris system will now auto reverse the remaining un-captured amount."

Voiding a capture or purchase requires a `purchasecorrection` command; however, this is not compatible with an authorization. So to allow both authorizations in addition to captures/purchases to be voided, you must pass :authorization => true in as an option to the `void` call. With no options, void will assume a capture or purchase as was the historical behaviour.

Note that voids on a capture/purchase can only be performed the same day so long as the batch that contains the original transaction remains open. From the Moneris documentation:

"A Void can be performed against a transaction as long as the batch that contains the original transaction remains open. When using the automated closing feature Batch Close occurs daily between 10 – 11 pm EST."