Mercury: Support card-less capture and refund
Mercury is first a POS system, and only second does online transactions.
Due to this, the standard "get back an authorization you can use to do
future operations on this transaction" flow must be explicitly enabled.

This enables that flow by having Mercury return a "OneTime" token for
every #authorize and #purchase which is stored in the authorization
string and can then be used to do captures and refunds. This also adds
the amount to the authorization string so that a capture can be run
without having to re-specify the amount.

This type of "one-time tokenization" is now the default for Mercury
since it matches so much better with expected ActiveMerchant semantics.
If the old behavior is desired, pass `tokenization: false` in the
options hash when creating the gateway.