SecureNet: Make interfaces more consistent
In particular, the #capture, #credit, and #void calls were inconsistent
with the other gateways in ActiveMerchant in that they took additional
parameters.  It's important for the gateways to share the same interface
and use the options hash for gateway specific stuff.

Now we use the authorization to hold the pieces SecureNet needs for
capture, credit, and void.