Braintree Blue: Scrub invalid emails and zips
Braintree is strict with emails and zips, which can cause unexpected
failures. This simply substitutes nil if a known-bad email or zip is
passed in so that the transaction doesn't fail due to the "bad" input.

This also improves the remote tests, including giving more guidance on
tests that require further configuration before they'll pass.