Bogus: Add Check support
Provides support for passing valid and invalid account numbers to the
bogus gateway and returning relevant error messages.

A valid account number is any account number that ends with 1.
An invalid account number is any account number that ends with 2.

Any object that responds to `account_number` will be treated as a check.

Note, because ActiveMerchant::Billing::Check has a `number` attribute
in addition to `account_number`, priority is given to `number` for
backwards compatibility.

CREDIT_ERROR_MESSAGE was removed because it was the same as
ERROR_MESSAGE.

Closes #944.