Balanced: Allow working with balanced.js
When using balanced.js, you get a card_uri without needing to use
`store`. This means that if you perform an operation with a card URI
and an email address, you get a "Funding instrument is not associated
with account" message, along with an error from Balanced.

By always associating the card to the (created) account if you're not
given an `account_uri`, you can avoid this issue and are able to use
balanced.js with the BalancedGateway.

Closes #992.