Balanced: Allow passing customer name
Since card tokens that are generated with balanced.js do not need to be
stored, allow `purchase` to accept a `name` option that will be set to
the user's account when doing a purchase with their newly tokenized
card. This saves the developer from doing an extra POST with the `store`
action.

Closes #995.