PlugNPay: Fix some remote tests
The PlugNPay test gateway now requires the name to have 'cardtest' for
the transaction to work.  This change gets some of the remote tests to
work again.

I'm still having difficulty with the remote tests for capture, void, and
refund.  Tech support for PlugNPay says that those operations are not
supported on their test gateway.  And they confirmed this fact on the
phone.

I'm hesitant to remove the remote tests for capture, void, and refund
though since they've been added in the past and they were clearly
working for someone at some point.  I don't know if PlugNPay adjusted
their test gateway to no longer support them.  Or if their tech
support is simply wrong and that there's actually some gateway setting
needed to enable those operations.  I'll leave those tests for now.