Wirecard: Handle a utf-8 description
If you specify a description that includes utf-8 characters, Wirecard
returns the following error:

Content of FunctionID is not according to the given content
restrictions. (1. Content of FunctionID must be alphanumerical with a
size of 0 to 32 characters.)

We now take a somewhat heavy handed approach with that description
field.  If Wirecard wants ascii, we'll give them ascii.  And we'll
replace non-ascii chacters with a question mark.