WePay: Fix for a full refund
From the WePay docs:
"This amount must be less than the 'net' of the transaction.
To perform a full refund, do not pass the amount parameter."

If you pass the full amount on a refund, WePay don't like that. So now
we only pass an amount if it's not equal to the original amount of the
purchase or capture.

Relevant WePay docs:

https://www.wepay.com/developer/reference/checkout#refund

This also fixes the interface for #capture.

Closes #1079.