Wepay: Fix for a full refund
From the Wepay docs:
"This amount must be less than the 'net' of the transaction.
To perform a full refund, do not pass the amount parameter."

If you pass the full amount on a refund, Wepay don't like that. So now
we only pass an amount if it's not equal to the original amount of the
purchase or capture.

Relevant Wepay docs:
https://www.wepay.com/developer/reference/checkout#refund