Merchant Warrior: Scrub names
Merchant Warrior is somewhat particular about what characters are
acceptable for the name on a card.  Their docs state:

"Only alphanumeric characters, hyphens, spaces and full stops are
allowed."

http://dox.merchantwarrior.com/files/MWE%20API.pdf

We now strip out the characters that Merchant Warrior doesn't like
because we don't want payments failing because the gateway isn't fond of
an & character.

Closes #1141.