Deprecate recurring functionality
There is too much variance between the recurring functionality provided
by gateways for ActiveMerchant to do a good job of abstracting it. Thus
this deprecates all such functionality with the intention of stripping
it out completely at some future point.

Note that this does *not* refer to tokenization support, but rather the
ability to set up and manage the ability some gateways support to
automatically charge a card on a recurring basis.