Elavon: Make void work for authorizations
The void operation we had worked for purchases but not for
authoriztions.  We were using a CCVOID operation.  Now we're using a
CCDELETE operation.

From the Elavon docs:
The `ccvoid` is a transaction that removes a Sale, Credit or Force
transaction from the open batch... The `ccvoid` command is typically
used for same day returns or to correct cashier mistakes.

The `ccdelete` is a transaction that deletes and attempts a reversal on a
Sale or Auth Only Credit transaction.

Closes #1261.