Stripe: Add ability to verify a card
For Stripe, a verification is an `authorization` followed by a `void` of
that authorization.

The minimum authorization Stripe allows is 50 cents so we use that.

Relevant docs:
https://stripe.com/docs/api#create_charge

We're now taking advantage of the ability of a MultiResponse to ignore
the result of some of the responses.  In this case, we don't want the
`verify` to fail if the `void` fails.  The important part is the
`authorize`.