Paypal: Add ability to verify a card
Paypal supports zero dollar authorizations, but only for mastercard and
visa.  Relevant docs:

https://developer.paypal.com/docs/classic/paypal-payments-pro/integration-guide/WPDPGettingStarted/#cardverifications

To verify a card, we do a zero dollar `auth` if it's mastercard or visa.
Otherwise, we do a $1 `auth` followed by a `void`.

We take advantage of the new ability in MultiResponse to ignore the
results of parts of the response.  In this case, if the void fails, we
don't care.