Authorize.net: Add ability to verify a card
I see some discussion out there on the interwebs around Authorize.net
supporting $0 authorizations for Visa only but it looks like support for
it is dependent on the merchant account sitting behind the gateway
account.

To make `verify` work for all authorize.net accounts, it's implemented
as a $1.00 `auth` followed by a `void`.

Closes #1281.