Vindicia: Stop using the vindicia-api gem
I was not able to run the remote tests due to not being able to get a
Vindicia dev account in a reasonable amount of time. If/when I get one
I'll run them, but would be great if anyone with credentials could try
them out (or provide them to me).

That said this ended up being quite straightforward as the gem wasn't
actually doing much.