Worldpay: Improve address defaulting
We've seen that Worldpay can, without warning, turn on the requirement
for address1 and city.  Previously, we were defaulting things like
postal_code and state.  Now we default more of the address fields if
they're not specified.

In theory, this should help some customers, it shouldn't hurt any
existing customers, and it should help customers down the road if
Worldpay changes the address requirements on an account.