eWay Rapid: Tweak authorization support
We now pass a "Method" of "Authorise" for an authorization as per this
documentation:

http://api-portal.anypoint.mulesoft.com/eway/api/eway-rapid-31-api/docs/reference/preauth#Authorisation

A customer reported that an authorize was showing up as a purchase.
This is an attempt to address that.  The sandbox doesn't seem to respond
differently for this "Method" change but I'm hopeful that production
will see a difference.  Either way, this brings us more inline with the
current eWay Rapid docs.