Cybersource: Add verify ability
I would have liked to have a remote test or two for verify.
Unfortunately, I can't seem to get the Cybersource remote tests to run.
The credentials I have are working, but I'm getting an error code of 150
(General system failure) in each of the tests.  It's maddening.  I'm not
sure if there's something amiss with our Cybersource sandbox account or
not.  I'm left to wonder if anyone can actually run them.