BraintreeBlue: Improve AVS handling
BraintreeBlue gives back an AVS street response and an AVS zip response.
We now map those responses to the standard codes found in AVSResult.

The approach to do the mapping is a bit verbose but hopefully easy to
understand.  It's quite similar to how the Stripe adapter is handling
its AVS mapping.

All of the possible return values for the 2 avs results are listed here:

https://developers.braintreepayments.com/ios+ruby/reference/objects/transaction#transaction.avs-postal-code-response-code

The test uses Braintree's sandbox test values:

https://www.braintreepayments.com/docs/ruby/reference/sandbox#avs_postal_code_response_codes