AuthorizeNet: Prevent test mode using live gateway
Authorize.net allows one to switch their production gateway account into
Test Mode.  If the account is in Test Mode, the response will include

<testRequest>1</testRequest>

Docs for it:

http://developer.authorize.net/api/reference/#payment-transactions-charge-a-credit-card

If the gateway instance you're dealing with is connecting to the
test_url, it matters not.  But if you're working with a live gateway
connecting to the live_url and you get back a response indicating that
the gateway backend is actually in Test Mode, then that's a serious
issue and the transaction should fail.  We don't want customers thinking
they're collecting real money when in fact they're not.