[POSSIBLE BREAKAGE] Stripe: Make version explicit
Stripe has many API versions, some of which are breaking changes:

https://stripe.com/docs/upgrades#api-changelog

If someone today created a new Stripe account, the remote tests would
not pass because Stripe has changed some of the attribute responses around
`cards` (used by our `store` method):

https://stripe.com/docs/upgrades#2015-02-18

In addition, the `statement_description` no longer works since it's been
replaced by `statement_descriptor`.  Clients of Active Merchant may be
using `statement_description` so we're now mapping that to
`statement_descriptor` under the covers.

So how do we solve these issues?  This commit makes the specification of
the Stripe API version explicit in the headers.  Previously, you could
specify the version you wanted to use for a particular transaction.  Or,
you could specify it when you instantiated the gateway.  In this commit,
we still allow such customizations.  If no version is specified though,
we explicitly use the current latest version (2015-04-07).

This improves things such that if Stripe releases a new, breaking API
version, clients using Active Merchant won't be broken.  If we want to
upgrade to the default version Active Merchant uses, we can do so
explicitly and change the underlying options we're passing accordingly.

If you're already specifying the Stripe API version, then this commit
shouldn't break you.  If you're relying on the default API version of
the underlying Stripe account, there may be issues which we'll fix as we
learn of them.