CenPOS: Simplify currency handling
CenPOS docs refer to using ISO 4217 (like most of the other gateways we
support).  No need to use the currency number and no need to maintain a
list which will be out of date if CenPOS adds more currencies.  Now we
simply let clients of ActiveMerchant pass in any currency and we pass it
straight through.