Stripe: Force USD for verify
The minimum authorization amount in Stripe is $0.50 USD.  Previously we
were sending 0.50 for the amount but then allowing clients of
ActiveMerchant to specify the currency in the options.  For some
currencies, this fails with a message such as:

There was a problem verifying your card details. Amount must convert to
at least 50 pence. $0.50 converts to approximately £0.32.

So now, we force the authorization used for `verify` to USD.