Fat Zebra: Fix refund and store signatures
The third parameter to refund needs to be an options hash to make it
consistent with the other gateways.  And the store method needs to take
an options hash as well.

Things would also blow up if an ip address and an order_id weren't
specified.  Now the adapter provides reasonable defaults for those
options.