Sage: Credit really is credit not refund
Revert "Sage gateway: Deprecate #credit call"

This reverts commit 425ffea0837ef592949d809735c855b647c22131.

I had deprecated the `credit` call 2 years ago thinking at the time that
the `credit` was really a refund.  In fact it was not.  It's a general
credit that isn't based on a previous transaction id.

I don't see an ability to do a standard `refund` in the Sage docs:

http://www.coderxo.com/docs/BankcardHTTPS.pdf

For now, this commit removes `refund` since the interface was broken.