Fix nil error on empty authorization result
Meh. I started to go down the path of forcing the authorization result to be an
empty string instead of nil, but then noticed that there are other values in the
response hash that can be nil. So, I opted for consistency and modified the code
to handle nils. I avoided the use of `try`, however.