NMI gateway (non-Auth.net mode) support
This is a rewrite of the NMI adapter to support their native API and not
the previous Auth.net emulation mode API. Every attempt has been made to
retain backwards compatibility. However, there may be small discrepancies
depending on your usage of the API.

What is known is that there are a few fields without a sensible mapping
to the new API. These request params include:

* cardholder_authentication_value and authentication_indicator which are
third party 3D Secure authentication fields. The NMI adapter doesn't
yet support these fields in its native API.
* solution_id (application_id), which identifies the agent making the API
call, has no equal in the native API.

Also, the native API does not support the notion of a fraud_review, so that
functionality is not available in the rewritten implementation.

The advantages of using the NMI native API include being on the most
current version which will make future development easier, as well as
support for features such as ACH/echeck payment methods (which this
rewrite does include).