Provide better insight to CVV usage in requests
Sometimes, the CVV isn't required for a transaction so it's helpful to
know in the transcript whether or not it was actually sent with a
request. With the current scrubbing logic, "[FILTERED]" will always be
used even if the CVV was not included. This replaces "[FILTERED]" with
"[BLANK]" when the CVV is nil, and empty string, or whitespace.

On the other hand, some gateways do not require a CVV on every request,
in those cases, a guard clause is being added to not add the CVV tag/param/key
to the request at all.

This PR includes changes for the following gateways, but more will be
added over time.

- AlliedWallet
- Paypal
- Openpay
- Redsys
- NMI