Micropayment: Allow project to be specified
In commit 2bdbccac2a8f9e63f1fa8f12316db191215e040, I saw a reference to
`sprdly` as a default value of the `project` parameter.  This in effect
meant that the project parameter was being passed for all transactions.
It's an optional parameter, and having `sprdly` as the default doesn't
seem right.

That being said, I shouldn't have removed that entire line though since
some customers do in fact want to specify the `project` for some
transactions.  This commit adds back that line, but without the default
value.