Stripe: Allow custom zero decimal currencies
The Stripe adapter has been overriding which currencies are zero decimal
currencies for awhile.  The parent Gateway class defines a default list
but Stripe has its own list:

https://support.stripe.com/questions/which-zero-decimal-currencies-does-stripe-support

The ability to override was broken in this commit:
https://github.com/activemerchant/active_merchant/commit/2a35543b4a8348cd115f4b2f138731e5f3c4a080

since the `non_fractional_currency?` method was changed to a class
method.

This commit allows us to override it again.  It also adds the VND
currency to Stripe's list since they now list that currency in their
docs.