TWD isn't a zero decimal currency
It was made a zero decimal currency in this commit:
https://github.com/activemerchant/active_merchant/commit/1d36d156ff3ec8b556082b3cf4566080769fc7d8

This was likely primarily for Paypal Express which _does_ list it as a
zero dollar currency:

https://developer.paypal.com/webapps/developer/docs/classic/api/currency_codes/

It's _not_ listed as a zero decimal currency in the ISO reference XML:

http://www.currency-iso.org/en/home/tables/table-a1.html

And Stripe has confirmed that it's not a zero decimal currency.

All that to say that by default, a Gateway that uses `localized_amount`
will no longer treat TWD as a zero decimal currency.  Our Paypal Express
adapter now specifies a custom list of zero decimal currencies that's in
line with their docs.