Cardstream: Improve default currency handling
If no currency is specified, use the gateway's default (GBP).  If a
currency is specified, use that currency.  If we don't support that
explicitly stated currency, then no longer default to GBP.  Instead,
the transaction should fail with a currency error.

Also added the BRL currency to the supported list.