FirstDataE4: Fix void for even dollar transactions
In 7e0508bcce8b76043c187851730480d7c2e272e3, void was fixed when the
amount of the auth was something like 29234 and `to_i` was giving an
amount that didn't match.  In that commit, we debated using `round` or
just removing the decimal point to avoid weird rounding errors.  What we
didn't realize was that when the amount is 5000 ($50.00), the gateway
returns back 50.0 in the response rather than 50.00.  So simply removing
the decimal point doesn't work in that case since the amounts won't
match.

This commit switches to `round` which may end up working all of the
time.  If it turns out to not work in some cases, we may end up taking
a different approach such as passing through the initial amount of the
auth all the way through to the `commit` method such that it can be
remembered rather than relying on the gateway's response body
amount.