Element: Use a better MotoECICode default
From Element support:

For Ecommerce transactions that are processed via HTTPS, the MotoECICode
should definitely be submitted as
7/NonAuthenticatedSecureECommerceTransaction to process correctly and
qualify at the correct interchange rate.  It's a secure ecommerce
transaction that is not using Verified by Visa or MasterCard SecureCode
functionality.  This is how all Ecommerce transactions on our platform
are submitted for processing.