Store action and authorization in the response
Keeping the action with the authorization means that subsequent requests
to the gateway using that authorization response can intelligently build
the correct call to the gateway.

The most direct example is the void action. Transaction Express has
multiple void types. Without storing the action in the authorization
field the end user would need to keep track of the action and choose the
correct void type.